docker-requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
dnspython==2.7.0
docopt==0.6.2
email-validator==2.0.0
Flask==2.3.3
Flask-WTF==1.1.1
fonttools==4.58.2
fpdf2==2.8.3
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
num2words==0.5.14
packaging==25.0
pillow==11.2.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
setuptools==80.9.0
six==1.17.0
tqdm==4.67.1
Werkzeug==2.3.7
wheel==0.45.1
WTForms==3.0.1
weasyprint==62.3
Flask-Caching==2.0.2
redis==4.6.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib